refactor(SystemConstant): 시스템 상수 자료형 변경
byte -> int

+++ src/main/java/baseball/SystemConstant.java
@@ -1,7 +1,7 @@
 package baseball;
 
 public class SystemConstant {
-    final byte MAXIMUM_DIGIT = 3;
-    final byte END_GAME = 2;
-    final byte NEW_GAME = 1;
+    final int MAXIMUM_DIGIT = 3;
+    final int END_GAME = 2;
+    final int NEW_GAME = 1;
 }
